minor updates
changed vertex vector to a short, added a secondary line check when x position is greater than or equal to the min of the two vector x positions, added a simple check to pass over any y positions that would be on or underneath the lava (could probably optimize this in normal check rather than position check)

+++ Magic.cpp
@@ -65,7 +65,7 @@ float dist_calc(vector<float> x, vector<float> y) {
 	return sqrtf(powf(y[0] - x[0], 2) + powf(y[1] - x[1], 2) + powf(y[2] - x[2], 2));
 }
 
-float line_point(vector<float> p1, vector<float> p2, float x, bool followY) {
+float line_point(vector<int32_t> p1, vector<int32_t> p2, float x, bool followY) {
 	if (followY) {
 		return (p2[1] - p1[1]) / (p2[0] - p1[0]) * (x - p1[0]) - p1[1];
 	}
