+++ Magic.h
@@ -12,7 +12,7 @@ using namespace std;
 pair<int16_t, float> calc_intended_yawmag(int8_t stick_x, int8_t stick_y);
 bool check_inbounds(struct Mario m);
 float dist_calc(vector<float> x, vector<float> y);
-float line_point(vector<float> p1, vector<float> p2, float x, bool yOrZ);
+float line_point(vector<int32_t> p1, vector<int32_t> p2, float x, bool yOrZ);
 
 #endif
